settings: fix incorrect data type for booleans
Python booleans are a subclass of integers, which means that
`isinstance(False, int) == True`. When creating (or updating) a setting,
we first check if a value is an integer then check if it is a boolean.
Because booleans are integers, we always treat booleans as ints. Checking
if a value is a boolean first will allow us to distinguish between the
types correctly.

Signed-off-by: Sam Friedman <sam@golioth.io>